move input to separate folder
