+++ src/components/Dashboard/Dashboard.css
@@ -19,6 +19,6 @@
 }
 
 .dashboard_cardsCont {
-    width: 80%;
+    width: 95%;
     
 }
