+++ src/components/Dashboard/Dashboard.js
@@ -43,6 +43,7 @@ return (
                     key={idx} 
                     boardTitle={board.title}
                     selectBoard={() => selectBoard(board.idx)}
+                    nestSeq={[board.idx]}
                     />))
                 :
                 null
